+++ TES.SDK.Examples/TesExamples.cs
@@ -5,6 +5,7 @@ using Azure.Identity;
 using Azure.Storage.Blobs;
 using Tes.Models;
 using Tes.SDK;
+using static Tes.SDK.TesClient;
 
 namespace TES.SDK.Examples
 {
@@ -53,7 +54,7 @@ namespace TES.SDK.Examples
 
             if (completedTask.State == TesState.COMPLETE)
             {
-                var client = new BlobClient(new Uri(task.Outputs.First().Url), new DefaultAzureCredential());
+                var client = new BlobClient(new Uri(task.Outputs.First().Url), new AzureCliCredential());
                 var downloadResponse = await client.DownloadContentAsync(CancellationToken.None);
                 var output = downloadResponse.Value.Content.ToString();
                 Console.WriteLine(output);
@@ -61,7 +62,7 @@ namespace TES.SDK.Examples
             else
             {
                 Console.WriteLine($"Failure reason: {completedTask.FailureReason}");
-                var logs = await tesClient.DownloadLogsAsync(completedTask, _storageAccountName, CancellationToken.None);
+                var logs = await DownloadLogsAsync(completedTask, _storageAccountName, CancellationToken.None);
 
                 foreach (var key in logs.Keys)
                 {
@@ -69,6 +70,35 @@ namespace TES.SDK.Examples
                     Console.WriteLine(logs[key]);
                 }
             }
+        }
+        public enum TesLogType { NotSet = 0, ExecStdOut, ExecStdErr, DownloadStdOut, DownloadStdErr };
+        public async Task<Dictionary<TesLogType, string>> DownloadLogsAsync(TesTask tesTask, string storageAccountName, CancellationToken cancellationToken)
+        {
+            Dictionary<TesLogType, string> logs = new Dictionary<TesLogType, string>();
+            var containerUri = new Uri($"https://{storageAccountName}.blob.core.windows.net/tes-internal");
+            var client = new BlobContainerClient(containerUri, new AzureCliCredential());
+            var blobs = client.GetBlobs(prefix: $"tasks/{tesTask.Id}").ToList();
+
+            foreach (var blob in blobs)
+            {
+                TesLogType logType = blob.Name switch
+                {
+                    var name when name.StartsWith("exec_stdout") => TesLogType.ExecStdOut,
+                    var name when name.StartsWith("exec_stderr") => TesLogType.ExecStdErr,
+                    var name when name.StartsWith("download_stdout") => TesLogType.DownloadStdOut,
+                    var name when name.StartsWith("download_stderr") => TesLogType.DownloadStdErr,
+                    _ => TesLogType.NotSet
+                };
+
+                if (logType != TesLogType.NotSet)
+                {
+                    var blobClient = client.GetBlobClient(blob.Name);
+                    var text = (await blobClient.DownloadContentAsync()).Value.Content.ToString();
+                    logs.Add(logType, text);
+                }
+            }
+
+            return logs;
         }
     }
 }
